+++ containers/sunshine/overlay/cloudy/bin/start-heroic.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+wait-x-availability.sh
+
+echo "Starting Heroic Games Launcher"
+
+# Start Heroic Games Launcher
+heroic