+++ src/providers/dummy/configurator.ts
@@ -2,7 +2,7 @@
 import { AnsibleConfigurator, AnsibleConfiguratorArgs } from '../../configurators/ansible';
 import { DummyInstanceStateV1 } from './state';
 
-export interface DummyConfiguratorArgs extends AnsibleConfiguratorArgs {}
+export interface DummyConfiguratorArgs extends AnsibleConfiguratorArgs<DummyInstanceStateV1> {}
 
 export class DummyConfigurator extends AnsibleConfigurator<DummyInstanceStateV1> {
 
@@ -12,6 +12,12 @@ export class DummyConfigurator extends AnsibleConfigurator<DummyInstanceStateV1>
 
     async doConfigure(): Promise<NonNullable<DummyInstanceStateV1['configuration']['output']>> {
         this.logger.debug(`Running dummy configurator for instance: ${this.args.instanceName}`)
+
+        if(this.args.provisionInput.configurationDelaySeconds && this.args.provisionInput.configurationDelaySeconds > 0){
+            const delay = this.args.provisionInput.configurationDelaySeconds * 1000
+            this.logger.debug(`Emulating configuration delay of Dummy instance ${this.args.instanceName}: ${delay}ms`)
+            await new Promise(resolve => setTimeout(resolve, delay))
+        }
 
         await super.doConfigure()